Improve the algorithm to match case when replacing.
This computes the Hamming distance between the original and its lower
case, upper case, and title case equivalents.  Whichever is the
closest match to the original, we do the same to the replacement.  The
lower case path also has a special check for whether the initial
letter is capitalized, so that sentence starts are correctly handled.